requirements.txt
# Web & API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0

# F1 Data & Processing
fastf1>=3.4.0
pandas>=2.2.3
polars>=1.0.0
# NumPy 2.1+ is required for Python 3.13 pre-built wheels
numpy>=2.1.0
pyarrow>=18.0.0

# Machine Learning & Stats
# These versions are needed for Python 3.13 compatibility
scikit-learn>=1.5.0
lightgbm>=4.5.0
pymc>=5.16.0
lifelines>=0.29.0
shap>=0.46.0

# Utils & Scrapers
beautifulsoup4>=4.12.3
requests>=2.32.0
httpx>=0.27.0
python-dotenv>=1.0.1
Pillow>=11.0.0
anthropic>=0.34.0
duckduckgo-search>=6.0.0

# Workflow & Cloud
prefect>=2.20.0
google-cloud-storage>=2.18.0
google-cloud-bigquery>=3.25.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
